% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/pad_election.R
\name{pad_election}
\alias{pad_election}
\title{Function to pad electoral data}
\usage{
pad_election(data, time, group, start, end)
}
\arguments{
\item{data}{A data frame with electoral data}

\item{time}{A vector of time data delimited by year (enclosed in "")}

\item{group}{A vector of country identifiers used to group data (enclosed in "")}

\item{start}{Numeric value for first year of time-series}

\item{end}{Numeric value for last year of time-series}
}
\description{
Function to pad electoral data
}

